Renamed struct source file and Added logic to test request
